emotion_project/requirements.txt
# requirements.txt (modifié)
tensorflow==2.17.0
pandas
numpy
scikit-learn
matplotlib
seaborn
neattext
nltk
nlpaug
joblib

gradio
fastapi
uvicorn[standard] # Le serveur pour faire tourner FastAPI
python-multipart  # Pour gérer l'upload de fichiers
Pillow # Pour la manipulation d'images
sentence-transformers